pyproject.toml
[project]
name = "devsecblueprint"
version = "1.0.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black==26.1.0",
    "boto3==1.42.49",
    "botocore==1.42.49",
    "certifi==2026.1.4",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "coverage==7.13.4",
    "cryptography==46.0.5",
    "ecdsa==0.19.1",
    "git-remote-codecommit>=1.17",
    "hypothesis==6.151.8",
    "idna==3.11",
    "iniconfig==2.3.0",
    "invoke>=2.2.1",
    "jinja2==3.1.6",
    "jmespath==1.1.0",
    "markupsafe==3.0.3",
    "moto==5.1.21",
    "mypy-extensions==1.1.0",
    "packaging==26.0",
    "pathspec==1.0.4",
    "platformdirs==4.9.2",
    "pluggy==1.6.0",
    "pyasn1==0.6.2",
    "pycparser==3.0",
    "pygments==2.19.2",
    "pytest>=9.0.2",
    "python-dateutil==2.9.0.post0",
    "python-jose==3.5.0",
    "pytokens==0.4.1",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "responses==0.25.8",
    "rsa==4.9.1",
    "s3transfer==0.16.0",
    "six==1.17.0",
    "sortedcontainers==2.4.0",
    "urllib3==2.6.3",
    "werkzeug==3.1.5",
    "xmltodict==1.0.3",
]
